chore: entity matcher with new fuzzy_matcher
relying on semantic search would not be sufficient as it might return
results that are semantically similar not not actually relevant.

search query on the book title field: "Science Fiction"
semantic search results: "The science of dune" even when there are records
with "Science Fiction" in the title.

a combination of semantic search and fuzzy matching should make up for
this